moved alert functionality code inside a different alert.ts file
